move process check command code into separate pkg
